[skip ci] rename stoke to coherency
