+++ africanus/model/coherency/__init__.py
@@ -1,0 +1,1 @@
+from .conversion import stokes_convert
