+++ africanus/model/coherency/cuda/__init__.py
@@ -1,0 +1,1 @@
+from africanus.model.coherency.cuda.conversion import stokes_convert
