--- africanus/model/stokes/__init__.py
@@ -1,1 +1,0 @@
-from .stokes_conversion import stokes_convert
