--- africanus/model/stokes/cuda/__init__.py
@@ -1,1 +1,0 @@
-from africanus.model.stokes.cuda.stokes_conversion import stokes_convert
